week 4

+++ week_3_session-2/function2.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
-float sum (int a, int b)
+int sum (int a, int b)
 {
-    int answer;
+    float answer;
     answer = a + b;
     return answer;
 }
@@ -12,5 +12,5 @@ int main ()
     int y = 2;
     float z = sum(5, y); // calling our new function
 
-    printf ("The sum of 5 and %d is %d\n", y, z);
+    printf ("The sum of 5 and %d is %f\n", y, z);
 }
